1- Intrusion Detection/requirements.txt
torch>=1.9.0
torchvision>=0.10.0
facenet-pytorch>=2.5.0
opencv-python>=4.5.0
numpy>=1.21.0
Pillow>=8.3.0
pyttsx3>=2.90
